refactor: Extract WorkspaceClient (final sub-client)
Fifth and final batch in the TodoistApi domain-decomposition series
(follows #561, #562, #563, #564). Migrates the 21 remaining workspace
methods onto the sub-client pattern.

- `src/clients/workspace-client.ts` — all workspace-domain endpoints:
  invitations (get/getAll/delete/accept/reject), join, logo upload,
  plan details, user listing + CRUD, workspace CRUD, members activity,
  user task listing, invite/update/remove users, active/archived
  project pagination.

`TodoistApi` gains one more private field and ctor instantiation. The
21 public methods collapse to one-line delegations. Public API, JSDoc,
and test files are unchanged — all 479 tests pass.

After this, only `getUser` and `sync` remain as directly-implemented
methods on `TodoistApi` — they make sense as the facade's core
cross-cutting entry points (identity + Sync API escape hatch) rather
than a one-method sub-client each.

`src/todoist-api.ts`: 2045 → 1709 lines (-336). Down from the original
3455 across the series — net -1746, split cleanly across 16 sub-client
files plus a thin facade.
